64bit整数で足りるものは long long に変更
* int は基本的に long long に変更
* GMP が過剰なものも long long に変更
* int で十分なものはそのまま(OGRなど)
* GMP が必要なものはそのまま

+++ prime/prime2.c
@@ -6,7 +6,7 @@ char    sieve[SIEVE_SIZE];
 
 int main( void )
 {
-    int n, i;
+    long long n, i;
 
     /*  ふるいを初期化する    */
     for( n = 2; n < SIEVE_SIZE; n++ )
@@ -23,7 +23,7 @@ int main( void )
     /*  ふるいで残った数は素数である    */
     for( n = 2; n < SIEVE_SIZE; n++ ) {
         if( sieve[n] == 1 )
-            printf( "%d\n", n );
+            printf( "%lld\n", n );
     }
 
     return( 0 );
